Append package identifier to rx and rx2 modules in order to prevent package name conflicts.
