+++ redux-core/src/test/java/me/tatarka/redux/TestMiddlewareTest.java
@@ -1,6 +1,7 @@
 package me.tatarka.redux;
 
 import me.tatarka.redux.middleware.TestMiddleware;
+
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
