+++ redux-rx2/src/main/java/me/tatarka/redux/rx2/CompletableDispatcher.java
@@ -1,7 +1,9 @@
-package me.tatarka.redux;
+package me.tatarka.redux.rx2;
 
 import io.reactivex.Completable;
 import io.reactivex.disposables.Disposable;
+
+import me.tatarka.redux.Dispatcher;
 
 public class CompletableDispatcher extends Dispatcher<Completable, Disposable> {
     @Override
