+++ redux-rx2/src/test/java/me/tatarka/redux/rx2/ObservableDispatcherTest.java
@@ -1,13 +1,19 @@
-package me.tatarka.redux;
+package me.tatarka.redux.rx2;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-import rx.Observable;
-import rx.Single;
-import rx.functions.Action0;
-import rx.observers.TestSubscriber;
+import io.reactivex.Completable;
+import io.reactivex.Maybe;
+import io.reactivex.Observable;
+import io.reactivex.Single;
+import io.reactivex.functions.Action;
+import io.reactivex.subscribers.TestSubscriber;
+
+import me.tatarka.redux.Dispatcher;
+import me.tatarka.redux.Reducer;
+import me.tatarka.redux.SimpleStore;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -22,18 +28,15 @@ public class ObservableDispatcherTest {
 
     @Test
     public void subscription_receives_initial_state() {
-        TestSubscriber<String> testSubscriber = new TestSubscriber<>();
         SimpleStore<String> store = new SimpleStore<>("test1");
-        ObserveStore.observable(store).subscribe(testSubscriber);
-
+        TestSubscriber<String> testSubscriber = ObserveStore.flowable(store).test();
         testSubscriber.assertValue("test1");
     }
 
     @Test
     public void subscription_receives_updated_state() {
-        TestSubscriber<String> testSubscriber = new TestSubscriber<>();
         SimpleStore<String> store = new SimpleStore<>("test1");
-        ObserveStore.observable(store).subscribe(testSubscriber);
+        TestSubscriber<String> testSubscriber = ObserveStore.flowable(store).test();
         store.setState("test2");
 
         testSubscriber.assertValues("test1", "test2");
@@ -41,9 +44,9 @@ public class ObservableDispatcherTest {
 
     @Test
     public void canceled_subscription_no_longer_receives_updates() {
-        TestSubscriber<String> testSubscriber = new TestSubscriber<>();
         SimpleStore<String> store = new SimpleStore<>("test1");
-        ObserveStore.observable(store).subscribe(testSubscriber).unsubscribe();
+        TestSubscriber<String> testSubscriber = ObserveStore.flowable(store).test();
+        testSubscriber.dispose();
         store.setState("test2");
 
         testSubscriber.assertValue("test1");
@@ -54,7 +57,6 @@ public class ObservableDispatcherTest {
         final int JOB_COUNT = 100;
 
         ExecutorService exec = Executors.newFixedThreadPool(JOB_COUNT);
-        TestSubscriber<String> testSubscriber = new TestSubscriber<>();
         final SimpleStore<String> store = new SimpleStore<>("test");
         final Dispatcher<String, String> dispatcher = Dispatcher.forStore(store, new Reducer<String, String>() {
             @Override
@@ -62,7 +64,7 @@ public class ObservableDispatcherTest {
                 return action;
             }
         });
-        ObserveStore.observable(store).subscribe(testSubscriber);
+        TestSubscriber<String> testSubscriber = ObserveStore.flowable(store).test();
 
         List<Callable<Void>> jobs = new ArrayList<>(JOB_COUNT);
         for (int i = 0; i < JOB_COUNT; i++) {
@@ -91,8 +93,7 @@ public class ObservableDispatcherTest {
         };
         SimpleStore<String> store = new SimpleStore<>("test1");
         ObservableDispatcher<String> dispatcher = new ObservableDispatcher<>(Dispatcher.forStore(store, reducer));
-        TestSubscriber<String> testSubscriber = new TestSubscriber<>();
-        ObserveStore.observable(store).subscribe(testSubscriber);
+        TestSubscriber<String> testSubscriber = ObserveStore.flowable(store).test();
         dispatcher.dispatch(Observable.just("test2"));
 
         testSubscriber.assertValues("test1", "test2");
@@ -108,8 +109,7 @@ public class ObservableDispatcherTest {
         };
         SimpleStore<String> store = new SimpleStore<>("test1");
         ObservableDispatcher<String> dispatcher = new ObservableDispatcher<>(Dispatcher.forStore(store, reducer));
-        TestSubscriber<String> testSubscriber = new TestSubscriber<>();
-        ObserveStore.observable(store).subscribe(testSubscriber);
+        TestSubscriber<String> testSubscriber = ObserveStore.flowable(store).test();
         dispatcher.dispatch(Observable.just("test2", "test3"));
 
         testSubscriber.assertValues("test1", "test2", "test3");
@@ -125,8 +125,7 @@ public class ObservableDispatcherTest {
         };
         SimpleStore<String> store = new SimpleStore<>("test1");
         SingleDispatcher<String> dispatcher = new SingleDispatcher<>(Dispatcher.forStore(store, reducer));
-        TestSubscriber<String> testSubscriber = new TestSubscriber<>();
-        ObserveStore.observable(store).subscribe(testSubscriber);
+        TestSubscriber<String> testSubscriber = ObserveStore.flowable(store).test();
         dispatcher.dispatch(Single.just("test2"));
 
         testSubscriber.assertValues("test1", "test2");
@@ -136,17 +135,46 @@ public class ObservableDispatcherTest {
     public void dispatch_completable() {
         SimpleStore<String> store = new SimpleStore<>("test1");
         CompletableDispatcher dispatcher = new CompletableDispatcher();
-        TestSubscriber<String> testSubscriber = new TestSubscriber<>();
-        ObserveStore.observable(store).subscribe(testSubscriber);
+        TestSubscriber<String> testSubscriber = ObserveStore.flowable(store).test();
         final boolean[] completableCalled = new boolean[1];
-        dispatcher.dispatch(rx.Completable.fromAction(new Action0() {
+        dispatcher.dispatch(Completable.fromAction(new Action() {
             @Override
-            public void call() {
+            public void run() throws Exception {
                 completableCalled[0] = true;
             }
         }));
 
         testSubscriber.assertValues("test1");
         assertTrue(completableCalled[0]);
+    }
+
+    @Test
+    public void dispatch_empty_maybe() {
+        Reducer<String, String> reducer = new Reducer<String, String>() {
+            @Override
+            public String reduce(String action, String state) {
+                return action;
+            }
+        };
+        SimpleStore<String> store = new SimpleStore<>("test1");
+        MaybeDispatcher<String> dispatcher = new MaybeDispatcher<>(Dispatcher.forStore(store, reducer));
+        TestSubscriber<String> testSubscriber = ObserveStore.flowable(store).test();
+        dispatcher.dispatch(Maybe.<String>empty());
+        testSubscriber.assertValues("test1");
+    }
+
+    @Test
+    public void dispatch_non_empty_maybe() {
+        Reducer<String, String> reducer = new Reducer<String, String>() {
+            @Override
+            public String reduce(String action, String state) {
+                return action;
+            }
+        };
+        SimpleStore<String> store = new SimpleStore<>("test1");
+        MaybeDispatcher<String> dispatcher = new MaybeDispatcher<>(Dispatcher.forStore(store, reducer));
+        TestSubscriber<String> testSubscriber = ObserveStore.flowable(store).test();
+        dispatcher.dispatch(Maybe.just("test2"));
+        testSubscriber.assertValues("test1", "test2");
     }
 }
